Toolbar: split categories into own popover, add inline search
Categories got their own popover+trigger on the toolbar with «Создать
категорию» at the bottom and folder-count badges that hide on hover.
Filters popover keeps only scope controls (company, deviation, status).
Toolbar also gains a collapsible search that auto-expands when ?q= is
present in the URL.
