+++ src/components/toolbar.test.tsx
@@ -1,4 +1,6 @@
 import { fireEvent, render, screen } from "@testing-library/react";
+import userEvent from "@testing-library/user-event";
+import { MemoryRouter } from "react-router";
 import { describe, expect, test, vi } from "vitest";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import type { FilterState, Folder, SortState } from "@/data/types";
@@ -14,6 +16,7 @@ function renderToolbar(
 		sort: SortState | null;
 		onFiltersChange: () => void;
 		onSort: () => void;
+		initialEntries: string[];
 	}> = {},
 ) {
 	const props = {
@@ -32,19 +35,35 @@ function renderToolbar(
 	};
 	return {
 		...render(
-			<TooltipProvider>
-				<Toolbar {...props} />
-			</TooltipProvider>,
+			<MemoryRouter initialEntries={overrides.initialEntries ?? ["/"]}>
+				<TooltipProvider>
+					<Toolbar {...props} />
+				</TooltipProvider>
+			</MemoryRouter>,
 		),
 		...props,
 	};
 }
 
 describe("Toolbar", () => {
-	test("does not render a search input (search lives in the global header)", () => {
+	test("renders collapsed search button; input is not in the DOM", () => {
 		renderToolbar();
-		expect(screen.queryByRole("button", { name: "Поиск" })).not.toBeInTheDocument();
+		expect(screen.getByRole("button", { name: "Поиск" })).toBeInTheDocument();
 		expect(screen.queryByPlaceholderText(/Поиск/)).not.toBeInTheDocument();
+	});
+
+	test("clicking search button expands it into an input", async () => {
+		const user = userEvent.setup();
+		renderToolbar();
+		await user.click(screen.getByRole("button", { name: "Поиск" }));
+		expect(screen.getByPlaceholderText("Поиск…")).toBeInTheDocument();
+	});
+
+	test("auto-expands when URL already has ?q= query", () => {
+		renderToolbar({ initialEntries: ["/?q=арматура"] });
+		const input = screen.getByPlaceholderText("Поиск…") as HTMLInputElement;
+		expect(input).toBeInTheDocument();
+		expect(input.value).toBe("арматура");
 	});
 
 	test("renders filter button", () => {
